src/pyproject.toml
[tool.poetry]
name = "OmniAgent"
version = "0.1.0"
description = ""
authors = ["Jowo Rinpoche <rinpoche@jowo.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
sse-starlette = "^1.6.5"
loguru = "^0.7.0"
toolz = "^0.12.0"
google-search-results = "^2.4.2"
psycopg = "^3.1.12"
aiohttp = "^3.8.6"
langchain = "0.2.6"
pydantic-settings = "^2.0.3"
gptcache = "^0.1.42"
asyncpg = "^0.28.0"
sqlalchemy = "^2.0.22"
websockets = "^12.0"
aiocache = "^0.12.2"
pytz = "^2023.3.post1"
ccxt = "^4.3.7"
beautifulsoup4 = "^4.12.3"
markdown = "^3.6"
sqlalchemy-utils = "^0.41.2"
retrying = "^1.3.4"
langchain-postgres = "0.0.9"
psycopg2-binary = "^2.9.9"
langchain-google-genai = "^1.0.4"
psycopg-binary = "^3.1.19"
langchain-core = ">=0.2.9,<0.3"
langgraph="0.1.1"
langchain-google-vertexai = "^1.0.6"
langchain-community = "^0.2.6"
langchain-openai = "^0.1.13"
chainlit = "^1.1.305"
greenlet = "^3.0.3"
covalent-api-sdk = "^1.0.2"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.1"
pre-commit = "^3.7.0"
mypy = "^1.10.0"

[tool.ruff]
line-length = 150
indent-width = 4
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

[tool.ruff.lint.mccabe]
max-complexity = 7

[tool.ruff.lint]
select = [
    "AIR",
    "B",
    "C90",
    "E",
    "F",
    "FLY",
    "FURB",
    "I",
    "N",
    "PERF",
    "RUF",
    "SIM",
    "UP",
    "W"
]
[tool.ruff.lint.per-file-ignores]
"__init__.py"= ["F401"]
"chainlit_app.py"= ["F811"]
"omniagent/db/database.py"= ["E261"]
"omniagent/service/chat.py"= ["C901"]
"omniagent/service/history.py"= ["E711"]
"omniagent/service/session.py"= ["C901","B904"]
"omniagent/agent/system_prompt.py" =["E501"]
"omniagent/dto/task.py"=["N815"]
"omniagent/experts/__init__.py"=["SIM117"]
"omniagent/experts/swap_expert.py"=["SIM117"]
"omniagent/db/models.py"=["N815"]


[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

[tool.mypy]
disallow_any_unimported = true
disallow_untyped_defs = false
no_implicit_optional = true
strict_equality = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_return_any = true
check_untyped_defs = true
show_error_codes = true
disable_error_code = ["arg-type","call-arg","union-attr","operator","unused-ignore","import-untyped","assignment","no-any-return","no-any-unimported","import-not-found","misc"]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
